requirements.txt

faiss-cpu==1.10.0
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.52
langchain-text-splitters==0.3.8
numpy==2.2.4
openai==1.74.1
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
PyMuPDF==1.25.5
pypdf==5.4.0
python-docx==1.1.2
python-dotenv==1.1.0
tqdm==4.67.1
unstructured==0.17.2
unstructured-client==0.32.3

